Break up production file, fix Karma paths
