+++ karma.conf.js
@@ -1,4 +1,4 @@
-var karmaConfig = require('./gulp/config').karma;
+var karmaConfig = require('./gulpfile.js/config').karma;
 
 module.exports = function(config) {
   config.set(karmaConfig);
